// apps/api/src/common/enums/role.enum.ts
export enum Role {
  SUPER_ADMIN = "SUPER_ADMIN",
  COMPANY_ADMIN = "COMPANY_ADMIN",
  SALES_PERSON = "SALES_PERSON",
  OPERATIONS_ADMIN = "OPERATIONS_ADMIN",
  HR_ADMIN = "HR_ADMIN",
  SALES = "SALES",
  PRINTING = "PRINTING",
  PRINT_OPS = "PRINT_OPS",
  HR = "HR",
  FINANCE = "FINANCE",
  SUPPORT = "SUPPORT",
  SCHOOL_ADMIN = "SCHOOL_ADMIN",
  SCHOOL_STAFF = "SCHOOL_STAFF",
  PARENT = "PARENT"
}